requirements.txt
pycryptodome
sockets
PyInstaller
Flasks
Flask
